A test script for when methods get called on tied variables.
I was particularly interested in Tied Scalars for use with Delayed::Failure. I've
determined localizing doesn't work because it's implemented by fetching the value, storing
undef to initialize, storing the changed value, then restoring the old value.

I had hoped it would bypass my class altogether, but this is probably better for the
majority of uses of tying scalars.