Add expressions and such to bring it up to parity with perl.
Still not as fast as I would like it to be, but I suppose since I'm specifying a dense
tree, it's going to end up with about 1.5^depth nodes that it evaluates for every
pixel. Some of which are sin/cos, so somewhat expensive. Depth of 15 takes beween 1/3 and
1 second on my virtual machine.

I disabled the clamp functions because, like the constants, they provide really cool fixed
point stuff, but if overused by the random generator, produce pretty awful pictures.

As the depth of the nodes goes up, the abuse increases so for higher depth it's better to
not have the functions.