Change parameters from Box<T> to T&.
A little better. Still have the wonky need to reference and dereference the inner
structure members pointlessly in the match construct, but I found the as_ref method which
means my functions no longer need to know the reference comes from a box, which was
something they shouldn't need to know. Whether that object is on the heap or stack, it's
still an object somewhere and the function shouldn't need to know that to work on it.

It seems to have a lower minimum run time, but it could have just been the random
generation giving me better functions than it had previously.