Complete serial go version.
Runs about 3-5x slower than the rust/ada versions for size 15. I could presumably
parallelize it easily in return for storing the whole image in memory and get it to about
the same speed, but I could also do that in Ada (and I believe rust does all it's
ownership stuff to make thready memory management easier, so it should be able to also).