Finish up prototype.
Allow it to range through the full state space (it's really not that large) in case we
have solutions of length > limit. If there are, they will show up in predecessor forms so
I don't think it's neccessary, but figured I could do it anyways since the queue is
dwindling by then and adding it takes like 1 more second of runtime.