Simple script that investigated the amount of space I needed to store the solutions.
After shoving the sequence into wolfram alpha, it gave me an algorithm for the direct
values, but not the sum (which I'm sure is also computable since this is in the same
family as triangular numbers). So I just hardcoded the sums into the Ada program and put
them into an array.