Have haskell do it's own printing.
When testing I just used a simple perl script because it was simpler than writing this
haskell when I didn't know if it was correct. I wanted to use printf like everything else,
and haskell has it, but I would have had to install the Text module and I didn't feel like
doing so for just this program.

I'm also somewhat surprised mapi wasn't in the standard library. I know it is in F#.