Start playing with a new puzzle game.
So far I've just turned the board into a graph. I had hoped the minimal spanning tree
would have something to do with the solution, and while it does provide a lower bound it
is trival to see in a 2x2 board of [rg][bg] that while the minimal spanning tree provides
distances of 1, you would have to take 2 moves to solve the puzzle. Luckily lower bounds
are all you need to have a valid A* search, so I can at least solve it heuristically.