Add the large board.
Results in complete death of my virtual machine as the program grows to consume all memory
at around a depth of 10 nodes. This means it attempts to swap, and everything with the
hard drive on this virutal machine is horrid.

So now I have something to play with for the A* solver. I may have to do something like a
medium node first, since I have the feeling my spanning tree approach won't give the best
focusing and will be overly optimistic.