Puzzle 20, incorrect. Reports 4.52 as the answer which is rejected.
This seems like a valid way to solve the problem, the 'hint' for the problem mentions
linearity of expectation, so I can take a look at that and see if I get a different answer
and whether this is some kind of rounding error or I'm missing something bigger. The
actual sum was 4.519999992 or something, so I doubt it's rounding.