Oldest mistake in the book. I transitioned to the wrong successor state.
Instead of Sunny 1 transitioning to Sunny 1, it transitioned to Sunny 2 (so I counted the
number of duplicate items in sequence instead of the count of sequence breaks).

Having the simple version was nice because it told me I failed starting at 3 maximum days,
so I was able to work just from that small dataset and doing it all by hand, I realized
that my probability was wrong. Missed it at first because all the final transition
probabilities are correct on Day 2, they're just in the wrong place so I had to start
calculating the expected values before I realized they were wrong.