Programs that test a method of comparing fractions.
Not really sure why I called these farey fractions except that they use the farey sequence
to generate them in order. But this was an idea I had at an American football game back in
September. If you have two fractions like 5/12 and 6/13, you know 6/13 is better because
it can be seen as the same as starting at 5/12 and then succeeding on 1/1 trials
afterwards.

This takes that to the conclusion of that idea, performing the operation again and again
until it gets something it easily knows how to compare and then reporting the result. It
was made to be a manual method of fraction compare, and it's of questionable value there
since it can be a lot of operations until you get something nice, but it does work because
of Medial fractions being between their hosts.