Upgrade prototype to be smarter about letters.
Namely after each assignment, I go through the list of words and shrink the list of
available letters to the set of letters that appear in those words.

So if my word list for a given spot is XYY = SEE | BEE, I know that the X is a S or B and
the Y is an E.

At the moment I'm not enforcing uniqueness, so it comes up with a number of spurious
solutions to the cipher, but it does complete the test set in a reasonable 20 seconds on
this virtual machine (with the first solution being pretty instantaneous).