Instead of binning by size, bin by word type.
This is a string where every letter has been mapped to the number that represents the
first position in the string it was seen.

So 'cats' would become '1234' and 'sees' would become '1221'.

This means that we are working from a smaller set of possible words since a string that
can make cats cannot make sees due to letter uniqueness.

This additionally allows us to remove the part of the code in make_re that created back
references.  Since we're now guaranteed those letters will match, we don't need to check
it. This both makes the regular expression faster (since it doesn't have to check) but
removes the creation of the seen and count hash tables.

Overall it has a marginal improvement in speed. The input step takes more work since word
type is more complex than word length, but searching is faster since the word lists are
shorter.