Enforce ascii and add a debug flag.
The code now only works in ascii, so tell perl about it so it can do whatever it feels
like to optimize for the limited range.

Use the environment flag DEBUG to turn on and off debug output. So if debug is desired, you can run it as:
DEBUG=1 perl prototype.pl < test.txt