Start of the ada version of the project.
It currently just reads in the possible word list and the cipher text.

It has one optimization over the perl prototype (of arguable use) which is that it only
creates the structure for the patterns that exist in the word list. This keeps the load
time to be essentially nothing because instead of having to create a map of 17k patterns
it only has to make the tens of patterns that actually show up in common use.

The next step of the project is to create the decipher module that's going to take the
enciphered words and return a mapping table from cipher letters to normal letters
(probably with a dummy value like '.' for letters that don't exist in the cipher
text).

Some parts of what I envision doing doesn't feel the cleanest, but that's what I get for
working in a low level language and wanting to use arrays instead of maps for small sets.

So for example the mapping hash in perl would instead just be an array of 'a' .. 'z' of
array (1 .. 26). That would be initialized to all the letters. As I make smaller sets,
that would then get lessened. Perhaps I should make it of records though so I can store
the size (since I would presumably be going through them via binary search and so
iterating through them to find the sentinal is wasted effort).

For the majority of their lifetime, all the assigned letters would be so small that linear
search is applicable (my rule of thumb is about 10), but the most expensive time would be
when the word list is large. Perhaps that means I should special case the mapping being
all 26 letters in the filter code since it will probably be all 26 or some small set
smaller than 5 letters.