Reversed condition logic in word filtering.
That was a simple bug. It's still not working, but at least the filtering is sane now,
where the first word immediately gets filtered down to 1 word from 26 (which seems more
likely than the earlier 26 going to 25 words). I'm guessing the new error is that the
other case is also reversed in logic, but we shall see.