Constrain_Letters was using ci (the character index used in the previous use) instead of i (the loop variable of the current loop).
This mean that it was always attempting to reduce the set of letters for the same
encrypted letter over and over. It looks like my previous copy down logic also didn't
work, so if the list should have ended up empty it ended up with 1 character. I rewrote it
to a more efficent write_head/read_head and solved that issue.

The logic is still partially broken. It finds the correct solution to the problem, but a
number of words are not fully constrained when I get to the solution part. I think that
means something isn't passing the list of letters updated properly to constrain letters.