Fix inclusion.
I forgot to exit the loop when I wrote the word to the inclusion list. This meant it just
wrote inclusions until it got to the end and then future words had nowhere to write.
(i.e instead of ending up with 1 8 9 13 0 ... 0 I would end up with 1 1 1 1 1 ... 1)

There is a strange slowdown when I rot1 encode it, which sounds to me like I should put
the guess ordering logic into place.