Sort the letters by word frequency, which removes the performance penalty for rot1 encoding the test input.
I imagine I'm paying a performance penalty for the fact that everything is copying around
24 bytes for every string used (so if I reduce a word list from 3000 words to 2000 words I
copy 2000 24 character words instead of just 2000 integers), but in release mode it
currently runs in .1 seconds on my machine.

For comparison, perl runs in just over 1 second and debug mode runs in about 3/4 of a
second.