Rework to return solutions instead of printing them to the console inline.
Start conversion to take both arguments from the command line. I need the test text to
come from a filename so I can seek to the start and convert the file for output instead of
just printing the possible words. It's a little bit of a false optimization to not just
read the whole file into a buffer since I'm storying all the words so my size is going to
be relatively similar to the file anyways, but it feels better this way.