Finish haskell solution.
Not really the greatest. I probably could have gotten a lot nicer solution a lot faster if
the interactive shell in haskell was better. For some reason it always wants to get
terminal statements on every line. Which means I can't test out parts of my code
interactively to debug issues.

So the only way to test my file interactively is to load the whole thing every time, which
means all intermediate variables I created are destroyed (because they depended on the
types from the file). So there's lots of repeated building of state that should just be
sitting in the terminal.

Performance is not great. About what perl is getting. There's one small performance glitch
in this, which is that it will attempt to constrain duplicate letters in words multiple
times. Most likely the real issue is that I'm not tickling the compiler the right way
somewhere. I don't really care enough to find what part makes it mad and fix it though, so
that's the end of the haskell portion of this exercise.

Next I think I'll play with an idea I had for generating c code and then running
it. Essentially specializing the code for the patterns. So that if I have a pattern like
ABBA, it will compile the check to just look at the first and second letter of the word
since the rest don't matter.

I suppose I can get something similar to this in the base code by changing all my stored
words to remove the duplicate letters, but then I have the disadvantage that debug output
would no longer be real words and generating the code is more interesting than doing a
secondary map.