C version operational.
Per usual when writing C code, even though I was essentially just translating ada, I still
got lost in segfault land. One was definitely my fault where I had a loop boundary that
did j++ instead of j. The other was a difference in opinion of what returning a vector
from a map should do. C++ decided that it should make a clone of the vector on the stack,
I figured since I didn't use it for anything except to get the address it would give me
the address of the vector in its own storage. Fixed by changing the map to store
wordList*. I was a little sad because nothing told me about my error, and even valgrind
was obtuse about it. I eventually just decided since I was having issues with it and I
figured it could go wrong that I could change the ownership and see what happened.

Since I'm not using the generator for actual code generation (just data structures), I'm
going to go ahead and remove that handling. Also since I already had valgrind handy I
checked for leaks and realized that I wasn't freeing the allocated word vectors when I was
done with the guess, so I'll have to go and test that on the ada version. Also I wonder if
I can run valgrind on an ada executable, it still goes through the same compiler chain
except the frontend.