Fix memory leaks in Ada version.
So valgrind does work on Ada. It seems to slow it down a lot more than the C version did,
but that could just be that the C version was faster due to various optimizations I added
into that version, so making the code 10x slower is just more noticeable.

The only real leak was the one I knew about, but I also took the time to clear out the
initial word list in the main loop.