Start cleaning up memory usage.
Two changes in here (It should have been two commits, but I didn't submit the first one
because in the middle of submitting it I realized a better way to do a small thing, so I
did that small thing and then moved on without committing).

First is to allocate all the memory for the words in one oversized block instead of 14k
small blocks. When I am done and know how much memory I actually needed, I then go back
through and update the pointers accordingly.

The second was to allocate the inclusion array manually instead of having the compiler
initialize it for me. This also has the benefit of removing the temporary hack I had to
put in place to get the compiler to allocate space for me. Because the outer array is no
longer fixed size, that requires a bit of extra storage for the pointers, but I more than
save it in not having to have NUM_WORDS empty integer slots in my array. I could have done
it a more canonical way that would have kept the static size, but then the words would
have been broken up from the numWords count.