Finish set uniqueness. Implement simulation code.
I do a global sort every time I start simulation, which isn't the best but I didn't want
to deal with find a priority queue that could handle re-inserting keys with new priorities
for the prototype. (The point being that it's better to simulate something with a few
possibilities and hope for a small improvement than to simulate something with 9! possible
choices.