Add two puzzles from the hardest difficulty setting in the phone app I was using for boards.
It was dissapointingly faster to do some of these boards than the earlier 'easier' boards
for some reason. I'm thinking of maybe making a visualizer for the board state as it makes
decisions. I haven't done anything like that in a while and it may be good to revive those
skills. Maybe something like a json driven html page. Have the debug messages I wrote turn
into structured messages and then use a stack of those to drive the page.