Make a json producing version.
Learned that the perl json module bases its decision on whether to output something as a
number or string on whether it was last USED in a numeric or string context. So something
like:

$data->{record_count} = @records;
print encode_json($data);
print "There were $data->{record_count} records";
print encode_json($data);

will result in the first record_count being numeric and the second being a
string. Interesting stuff. I knew perl kept track of the availability of a string or
numeric representation, but it's interesting that you can tell which was last used.