Some simple AMPL models for item acquisition.
I saw an article today which did something like this by hand on a GPU. This is a
simplified version of their model (but their model was poorly specified in the article
because it was providing shipping costs for things like pairs of items, but if an item was
in two pairs I'm not sure what the cost would have been).  If I play with this more, the
next step would be a weight based shipping cost. Then a weight category shipping cost
(Which is only harder because I have to do more min/max constraining to get it to bin
properly).

In reality, shipping is more determined by package size than weight to my knowledge, but
3D bin packing is a whole seperate optimization problem.