Day three puzzle.
This one actually required some thought. I knew the basic structure would be finding the
corners of the square it is in and then using that to determine the number of steps.

There was a bit of annoyance because the odd squares are the ones that complete a shell so
I had to dance around that point a little bit. Then it was just adding the number of steps
to get to that shell to the number of steps to a midpoint of the shell. I get the distance
to the midpoint by determining the distance to either the top right or bottom left corner
and comparing that to the shell number, which is also the distance from one of those
corners to the midpoint.