Write up and perform analysis.
It looks like the last analysis was just too far back in the past for me to see it in my
glance of the git logs, but I do see the norepeats.c, so I'll try and reproduce what it
did and see if it comes out with some mistakes. That would clear my conscience at least.