Letter filtering for word squares.
So the concept here is that we can determine if words are subsets of other words and we
would always be better off picking the other word. For example using were is strictly
worse than using wire. If two words have the same set of letters, we pick the shorter one,
so we want to use read in preference to reread.